requirements.txt
python-docx==1.1.2
openpyxl==3.2.0b1
gevent==24.2.1
requests==2.32.3
quart==0.19.6
hypercorn==0.17.3
chromadb==0.5.5
python-dotenv==1.0.1
psycopg==3.2.1
tiktoken==0.7.0
openai==1.45.1
pydantic==2.8.2
pandas==2.2.2
pdfminer==20191125
numpy==1.26.4
tldextract==5.1.2
selenium==4.24.0
webdriver-manager~=4.0.2
faircore==5.0.2
aiohttp==3.10.5
torch==2.4.1
transformers==4.44.2
datasets~=3.0.0
redis~=5.0.8
flask~=3.0.3
pillow~=10.4.0
weasyprint~=62.3
rdflib==6.3.2
pytesseract==0.3.13
validators==0.34.0
pathlib==1.0.1
fastapi==0.112.2
langchain==0.3.2
langchain-community==0.3.1
langchain-core==0.3.8
Colbert==0.40
huggingface-hub==0.24.6
SQLAlchemy==2.0.35
peewee==3.17.6
PyYAML==6.0.2
starlette==0.38.2
jwt==1.3.1
passlib==1.7.4
Markdown==3.7
docx2txt==0.8
rank-bm25==0.2.2
jsonlines==4.0.0
unstructured==0.15.13
tqdm==4.66.5
pypdf2==3.0.1
textract==1.6.5
pdf2image==1.17.0
textract==1.6.5
pdfminer.six>=20191110
argcomplete==1.10.0
beautifulsoup4==4.8.0
chardet==3.0.4
EbookLib==0.17.1
lxml==4.6.3
python-pptx==0.6.18
six==1.12.0
SpeechRecognition==3.8.1
xlrd==1.2.0